🔨 refactor: ts 파일로 수정
