+++ .github/workflows/pull_request.yml
@@ -22,6 +22,9 @@ jobs:
       - name: Install dependencies
         run: npm install
 
+      - name: Install ts-node
+        run: npm install -g ts-node typescript @types/node
+
       - name: Run Slack Notification Script
         env:
           SLACK_BOT_TOKEN: ${{ secrets.SLACK_BOT_TOKEN }}
@@ -29,4 +32,4 @@ jobs:
           SLACK_WEBHOOK_URL: ${{ secrets.SLACK_WEBHOOK_URL }}
           GITHUB_EVENT_PATH: ${{ github.event_path }}
           GITHUB_EVENT_NAME: ${{ github.event_name }}
-        run: node script/sendMessageToSlack.js
+        run: ts-node script/sendMessageToSlack.ts
