--- script/constant.js
@@ -1,6 +1,0 @@
-/* eslint-disable no-undef */
-const users = {
-  bicochan: 'U07C75ZKHEX',
-};
-
-module.exports = users;
