+++ script/constant.ts
@@ -1,0 +1,5 @@
+const users = {
+  bicochan: 'U07C75ZKHEX',
+};
+
+export default users;
